Building as executable
Changed and adjusted most of the project to build it into an executable.
fix a few problem that came up already but still a lot to fix to get everything back up and running.
Added the Option to Save Load or create new Projects.
Intensity Slider now with function
clip loading gets handled by a new worker to free up the ui worker.
Some CSS Adjustments for the clip-deck on smaller screen sizes
Fixed also a problem in ilda file parsing with format code 2  to assume 3 bytes instead of 8 wich caused malformed frames